requirements.txt
pandas==2.1.4
firebase-admin==6.4.0
python-dotenv==1.0.0
pymysql==1.1.0
paramiko==3.3.1
sshtunnel==0.4.2

